+++ pkg/resources/certificates.go
@@ -1,0 +1,65 @@
+//
+// Copyright 2020 IBM Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+package resources
+
+import (
+	"reflect"
+
+	operatorv1alpha1 "github.com/ibm/ibm-auditlogging-operator/pkg/apis/operator/v1alpha1"
+	certmgr "github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+// BuildCertsForAuditLogging returns a Certificate object
+func BuildCertsForAuditLogging(instance *operatorv1alpha1.AuditLogging, issuer string, name string) *certmgr.Certificate {
+	metaLabels := LabelsForMetadata(FluentdName)
+	var clusterIssuer string
+	if issuer != "" {
+		clusterIssuer = issuer
+	} else {
+		clusterIssuer = defaultClusterIssuer
+	}
+
+	certificate := &certmgr.Certificate{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: InstanceNamespace,
+			Labels:    metaLabels,
+		},
+		Spec: certmgr.CertificateSpec{
+			CommonName: name,
+			IssuerRef: certmgr.ObjectReference{
+				Name: clusterIssuer,
+				Kind: certmgr.ClusterIssuerKind,
+			},
+		},
+	}
+
+	if name == AuditLoggingHTTPSCertName {
+		certificate.Spec.SecretName = AuditLoggingServerCertSecName
+		certificate.Spec.DNSNames = []string{AuditLoggingComponentName}
+	} else {
+		certificate.Spec.SecretName = AuditLoggingClientCertSecName
+	}
+
+	return certificate
+}
+
+// EqualCerts returns a Boolean
+func EqualCerts(expected *certmgr.Certificate, found *certmgr.Certificate) bool {
+	return !reflect.DeepEqual(found.Spec, expected.Spec)
+}
